# json performance tuning