@xerial 2019-07-13
Thanks for the improvement! Now we can close #381